Replacing Earlier Subtree Metadata
Earlier, I was hand-creating subtrees using the "subtree-merge
strategy."  Since I was doing this in a clone of a GitHub repo, wanted
some way to preserve the subtree info, so that I could restore it from
any later clones.

I've since discovered the new "git subtree" feature.  [I'm on Ubuntu
12.04LTS.]

I still want to store information about my subtrees so that I can
reconfigure any future clones of the remote GitHub repo containing them.
So, out with the old, kinda-kludgy subtree-metadata, and in with the
new, cleaner form!